Restructure database again to use compound keys
Use compound keys for images and characters as a way to prevent duplicates when reloading. This effectively moved the FK relationship direction from the parent to the child.
i.e. when getting the image for a character you would search for images that have the compound key for that set and that character. Rather than using the UUID string stored in the character as the key for the images database